+++ toolbox/assets.mjs
@@ -15,7 +15,9 @@ export function createAssetContext(workingDirectory = root) {
       'css/syutoi.min': 'src/styles/main.css',
       'js/lightbox.min': 'src/client/lightbox.ts',
       'js/photoswipe.min': 'src/client/photoswipe.ts',
-      'css/lightbox.min': 'src/styles/lightbox.css'
+      'css/lightbox.min': 'src/styles/lightbox.css',
+      'js/search.min': 'src/client/search.ts',
+      'css/search.min': 'src/styles/search.css'
     },
     outdir: 'source',
     bundle: true,